working oc version TODO check work on the server

+++ core/game.py
@@ -152,7 +152,7 @@ class GameHistory:
 
     def zero_obs(self):
         # return a zero frame
-        return [np.zeros(self.zero_obs_shape, dtype=np.uint8) for _ in range(self.stacked_observations)]
+        return [np.zeros(self.zero_obs_shape, dtype=np.float32) for _ in range(self.stacked_observations)]
 
     def step_obs(self):
         # return an observation of correct format for model inference
